chore: move validation to another file
